+++ server/src/server.ts
@@ -3,17 +3,18 @@ import { deployMultisigWallet } from '../deployMultiSig';
 import cors from 'cors'; // Import cors module
 
 const app = express();
+app.use(bodyParser.json());
 const port = 8080;
 import * as dotenv from 'dotenv';
+import bodyParser from 'body-parser';
 dotenv.config();
 app.use(cors({
     origin: 'http://localhost:3000',
     credentials: true
 })); // Enable CORS
 
-app.post('/', async (req, res) => {
-    const {pk} = req.body
-    console.log('called')
+app.post('/create-multisig', async (req, res) => {
+    const { pk } = req.body;
     const { account, keys, accountContract } = await deployMultisigWallet(pk);
     res.json({ account, keys, accountContract });
 });